copy paste filesystem functions into new file
actual code unchanged, just doing this now for diff reasons
